test(router): move regression tests into mapped test files
